backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
werkzeug==3.0.1
bcrypt==4.1.2
redis==5.0.1
celery==5.3.4
openai==1.6.1
anthropic==0.8.1
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
pydantic==2.5.3
email-validator==2.1.0
validators==0.22.0
better-profanity==0.7.0
langdetect==1.0.9
requests==2.31.0
gunicorn==21.2.0
